Flyttar validering av respons opp i interfacet, så default-implementasjonen av den er tilgjengeleg for konsumentane som ønskjer det, og gir dei i tillegg moglegheita til å leggje på eiga validering sjølv
